requirements.txt
psycopg2~=2.9.6
graphene~=3.2.2
Flask~=2.3.2
python-dotenv~=1.0.0